split canvas.hpp
